+++ src/me/anno/ecs/components/physics/CustomBulletCollider.kt
@@ -1,7 +1,7 @@
 package me.anno.ecs.components.physics
 
-import org.joml.Vector3d
+import org.joml.Vector3f
 
 interface CustomBulletCollider {
-    fun createBulletCollider(scale: Vector3d): Any
+    fun createBulletCollider(scale: Vector3f): Any
 }
