+++ src/me/anno/ecs/components/physics/ScaledBody.kt
@@ -1,11 +1,12 @@
 package me.anno.ecs.components.physics
 
 import org.joml.Vector3d
+import org.joml.Vector3f
 
 class ScaledBody<InternalRigidbody, ExternalRigidbody>(
     val internal: InternalRigidbody,
     val external: ExternalRigidbody,
-    val scale: Vector3d,
+    val scale: Vector3f,
     val centerOfMass: Vector3d
 ) {
     operator fun component1() = internal
